docs: Improve developer onboarding
Without this, ppl will be lost in contributing.

By working on this page, we follow diataxis model:
- How to contribute is in CONTRIBUTING.md
- The learning experience of the contributor life is in journey
- The policy document explains why we have our policies.
  It was kinda a duplicate all over the place between processes
  and GOVERNANCE and CONTRIBUTING. Now it is split and linked.
- Reference documentation (the WHAT we do) are in different pages

Next to this, deprecations were all over the place.

This was hard to grasp. This is a problem, as one might not read the docs and badly
deprecate features.

This is solved by:
- Removing the duplicate content
- Splitting the content into 3 parts:
  - The DEPRECATING.md file containing the HOWTO deprecate (list of instructions)
  - The deprecation_policy.md containing WHAT do we consider a deprecation
  - An update to our policy.md to explain how our deprecations/removals are
    a special case of impactful changes.

Finally, security process where not present in the main docs.
It could be easily forgotten.

At the same time, the SECURITY.md is a well known expected location.
I did not want to lose that information or duplicate it.

I fixed it by adding a dynamic link between our docs and the usual
SECURITY.md location.

All the other _custom_ locations where either removed or
dynamically linked.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>